Use server-rendered HTML fields
We have a bug in the rendering of expandContent, where we overzealously
strip out some tags (notably del). We want to move away from
client-rendered HTML in general, favoring HTML fields provided by the
server.  By switching to the correctly-rendered API fields, we can avoid
the rendering bug and remove some dependencies, including the 3rd-party
library that led to the original rendering bug.